update listener api to return provided resources
in local.js there is a provides key that specifies what
the listener provides.  we can read this to know what
methods to call during init or client creation for functional tests.
additionaly the api endpoints needed better naming like 'ios' instead of
liteserv which can be confused with android-literve